refactor: mvc 패키지 구성
mvc 패키지 구성
