+++ src/main/java/race/Main.java
@@ -1,5 +1,10 @@
 package race;
 
+import race.race.CarRace;
+import race.race.TryCount;
+import race.view.InputView;
+import race.view.ResultView;
+
 public class Main {
 
     public static void main(String[] args) {
