+++ src/main/java/race/car/Car.java
@@ -1,4 +1,6 @@
-package race;
+package race.car;
+
+import race.race.RandomNumber;
 
 public class Car {
