+++ src/main/java/race/car/CarName.java
@@ -1,4 +1,4 @@
-package race;
+package race.car;
 
 import java.util.Objects;
