+++ src/test/java/race/CarNameTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
+import race.car.CarName;
+
 class CarNameTest {
 
     @Test
