+++ src/test/java/race/PositionTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
+import race.car.Position;
+
 class PositionTest {
 
     @Test
