+++ src/test/java/race/RandomNumberTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
+import race.race.RandomNumber;
+
 class RandomNumberTest {
 
     @Test
